move test to async2 and cleanup tests
